🐛 [Frontend] Fix: New project from menu (#7377)
