+++ services/static-webserver/client/source/class/osparc/dashboard/ServiceBrowser.js
@@ -32,16 +32,6 @@ qx.Class.define("osparc.dashboard.ServiceBrowser", {
     this.__sortBy = osparc.service.SortServicesButtons.DefaultSorting;
   },
 
-  properties: {
-    multiSelection: {
-      check: "Boolean",
-      init: false,
-      nullable: false,
-      event: "changeMultiSelection",
-      apply: "__applyMultiSelection"
-    }
-  },
-
   members: {
     __sortBy: null,
 
@@ -111,35 +101,6 @@ qx.Class.define("osparc.dashboard.ServiceBrowser", {
       const serviceData = card.getResourceData();
       this._openResourceDetails(serviceData);
       this.resetSelection();
-    },
-
-    _createStudyFromService: function(key, version) {
-      if (!this._checkLoggedIn()) {
-        return;
-      }
-
-      const studyAlias = osparc.product.Utils.getStudyAlias({firstUpperCase: true});
-      this._showLoadingPage(this.tr("Creating ") + studyAlias);
-
-      osparc.study.Utils.createStudyFromService(key, version)
-        .then(studyId => {
-          const openCB = () => this._hideLoadingPage();
-          const cancelCB = () => {
-            this._hideLoadingPage();
-            const params = {
-              url: {
-                studyId
-              }
-            };
-            osparc.data.Resources.fetch("studies", "delete", params);
-          };
-          const isStudyCreation = true;
-          this._startStudyById(studyId, openCB, cancelCB, isStudyCreation);
-        })
-        .catch(err => {
-          this._hideLoadingPage();
-          osparc.FlashMessenger.logError(err);
-        });
     },
 
     // LAYOUT //
